+++ example/lib/screens/others.dart
@@ -20,6 +20,32 @@ class _OthersState extends State<Others> {
 
   bool checked = false;
 
+  final items = [
+    TreeViewItem(
+      content: const Text('Work Documents'),
+      children: [
+        TreeViewItem(content: const Text('XYZ Functional Spec')),
+        TreeViewItem(content: const Text('Feature Schedule')),
+        TreeViewItem(content: const Text('Overall Project Plan')),
+        TreeViewItem(content: const Text('Feature Resources Allocation')),
+      ],
+    ),
+    TreeViewItem(
+      content: const Text('Personal Documents'),
+      children: [
+        TreeViewItem(
+          content: const Text('Home Remodel'),
+          children: [
+            TreeViewItem(content: const Text('Contractor Contact Info')),
+            TreeViewItem(content: const Text('Paint Color Scheme')),
+            TreeViewItem(content: const Text('Flooring weedgrain type')),
+            TreeViewItem(content: const Text('Kitchen cabinet style')),
+          ],
+        ),
+      ],
+    ),
+  ];
+
   @override
   void dispose() {
     flyoutController.dispose();
@@ -252,6 +278,59 @@ class _OthersState extends State<Others> {
                   ]),
                 ),
               ),
+            ),
+          ),
+          const SizedBox(height: 20.0),
+          SizedBox(
+            height: 380.0,
+            child: ListView(
+              scrollDirection: Axis.horizontal,
+              children: [
+                Container(
+                  constraints: const BoxConstraints(
+                    maxHeight: 380,
+                    minHeight: 380,
+                    maxWidth: 350,
+                  ),
+                  decoration: BoxDecoration(
+                    border: Border.all(
+                        color: FluentTheme.of(context).inactiveColor),
+                  ),
+                  child: TreeView(items: items),
+                ),
+                Container(
+                  constraints: const BoxConstraints(
+                    maxHeight: 380,
+                    minHeight: 380,
+                    maxWidth: 350,
+                  ),
+                  decoration: BoxDecoration(
+                    border: Border.all(
+                        color: FluentTheme.of(context).inactiveColor),
+                  ),
+                  child: TreeView(
+                    selectionMode: TreeViewSelectionMode.single,
+                    items: items,
+                    onItemInvoked: (item) => debugPrint('$item'),
+                  ),
+                ),
+                Container(
+                  constraints: const BoxConstraints(
+                    maxHeight: 380,
+                    minHeight: 380,
+                    maxWidth: 350,
+                  ),
+                  decoration: BoxDecoration(
+                    border: Border.all(
+                        color: FluentTheme.of(context).inactiveColor),
+                  ),
+                  child: TreeView(
+                    selectionMode: TreeViewSelectionMode.multiple,
+                    items: items,
+                    onItemInvoked: (item) => debugPrint('$item'),
+                  ),
+                ),
+              ],
             ),
           ),
         ],