+++ lib/fluent_ui.dart
@@ -56,6 +56,7 @@ export 'src/controls/inputs/slider.dart';
 
 export 'src/controls/navigation/navigation_view/view.dart';
 export 'src/controls/navigation/tab_view.dart';
+export 'src/controls/navigation/tree_view.dart';
 
 export 'src/controls/surfaces/calendar/calendar_view.dart';
 export 'src/controls/surfaces/bottom_sheet.dart';